+++ pkg/dds/bin/dds.dart
@@ -128,6 +128,7 @@ ${argParser.usage}
           'uri': dtdInfo.uri,
         },
     }));
+    stderr.close();
   } catch (e, st) {
     writeErrorResponse(e, st);
   }